core: merge ScheduleMetadataExtractors
just move "V1" into V2
